Added a footer for feedbacks, switched to friends instead of password on loading profile page and added confirmation before removing a friend or deleting a recommendation.

+++ src/client/profile_screen/components/friends.js
@@ -93,6 +93,8 @@ export function Friends() {
 
 
     const deleteFriend = (friend, index) => {
+        if (!window.confirm(`האם את/ה בטוח/ה שברצונך להסיר את ${friend} זה מרשימת החברים? `))
+            return;
         axios.delete("https://rbfserver.herokuapp.com/users/friends",
             {
                 headers: { Authorization: `Bearer ${token}` },
